requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
requests==2.32.2
torch==2.2.2
numpy==1.26.4
matplotlib==3.9.0
torchvision==0.17.2
scikit-learn==1.5.0
pillow==10.3.0
virtualenv==20.26.2